t210: orca: dts: Populate edp2lvds i2c bus
Set edp-lvds-i2c-bus-no property in E2606 dtsi.
Based on the value of this property, I2C1 or I2C4
can be used. If I2C1 to be used, set this property
to value 0, if I2C4 to be used, set this property
to value 3. Setting by default to 0. Accordingly
do the jumper setting j25 on the base board m3401.

Bug 200264272

Change-Id: I7492edb0bccf6d090138c55098bc458b03c96cc0
Signed-off-by: Gaurav Singh <gaursingh@nvidia.com>
Reviewed-on: http://git-master/r/1286674
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bibek Basu <bbasu@nvidia.com>